% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleShiftRangeWrapper.R
\name{double.shift.range}
\alias{double.shift.range}
\title{Double Shift Estimator Across Delta Range -- not right yet}
\usage{
double.shift.range(y, a, z, x, delta1, delta2, Y.est, A.est, Z.est,
  nfolds = 2, zmax = Inf, zmin = -Inf, alpha = 0.05, nbs = 10000,
  pos.cutoff = 500, ...)
}
\arguments{
\item{y}{your outcome variable (a vector)}

\item{a}{your treatment variable (a vector)}

\item{z}{your instrument (a vector)}

\item{x}{a dataframe of covariates}

\item{Y.est}{an algorithm for estimating the means of your outcome.
Should be 'glm', 'superlearner' or 'ranger'. If you choose superlearner,
you can also specify the libraries you would like to use. Default libraries
are c("SL.glm","SL.randomForest","SL.polymars","SL.mean").}

\item{A.est}{an algorithm for estimating the means of your treatment.
Should be 'glm', 'superlearner' or 'ranger'. If you choose superlearner,
you can also specify the libraries you would like to use. Default libraries
are c("SL.glm","SL.randomForest","SL.polymars","SL.mean").}

\item{Z.est}{an algorithm for estimating the means of your treatment.
Should be 'glm' in which case a glm is used to estimte the mean and variance
and a kernel is used to estimate the density, or 'flexcode'. If you choose
'flexcode', you can specify the regression function. The default is regressionFunction.NW}

\item{nfolds}{number of folds. Defaults to 2}

\item{zmax}{the upper bound on Z, default is Inf}

\item{zmin}{the lower bound on Z, default is -Inf}

\item{alpha}{the alpha level for your multiplier bootstrap confidence bands (default 0.05)}

\item{nbs}{the number of boostrap samples to take. Default is 10,000.}

\item{pos.cutoff}{the level at which you want to truncate pi(Z+delta)/pi(Z) for positivity
restrictions. Default is 50.}

\item{delta}{a vector of shift levels}
}
\value{
a list including an estimate of the effect for each delta value and its standard deviation,
as well as upper and lower confidence intervals for the pointwise and uniform case.
}
\description{
Estimates the effects of two shift amounts
across a range of values. Includes the multiplier bootstrap
for calculations of uniform confidence bands
}
